Address devin review comments
