+++ apps/webapp/app/routes/api.v1.tasks.$taskId.trigger.ts
@@ -19,17 +19,9 @@ import {
   handleRequestIdempotency,
   saveRequestIdempotency,
 } from "~/utils/requestIdempotency.server";
+import { sanitizeTriggerSource } from "~/utils/triggerSource";
 import { ServiceValidationError } from "~/v3/services/baseService.server";
 import { OutOfEntitlementError, TriggerTaskService } from "~/v3/services/triggerTask.server";
-
-const ALLOWED_TRIGGER_SOURCES = new Set(["sdk", "cli", "mcp"]);
-
-export function sanitizeTriggerSource(value: string | null | undefined): string | undefined {
-  if (value && ALLOWED_TRIGGER_SOURCES.has(value)) {
-    return value;
-  }
-  return undefined;
-}
 
 const ParamsSchema = z.object({
   taskId: z.string(),
