+++ apps/webapp/app/v3/runEngineHandlers.server.ts
@@ -750,7 +750,7 @@ export function setupBatchQueueCallbacks() {
               batchIndex: itemIndex,
               realtimeStreamsVersion: meta.realtimeStreamsVersion,
               planType: meta.planType,
-              triggerSource: meta.parentRunId ? "sdk" : "api",
+              triggerSource: meta.parentRunId ? "sdk" : meta.triggerSource ?? "api",
               triggerAction: "trigger",
             },
             "V2"
